// src/types/user.interface.ts
export interface IGeo {
    lat: string;
    lng: string;
};

export interface IAddress {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: IGeo;
};

export interface ICompany {
    name: string;
    catchPhrase: string;
    bs: string;
};

export interface IApiUser {
    id: number;
    name: string;
    username: string;
    email: string;
    address: IAddress;
    phone: string;
    website: string;
    company: ICompany;
};

export interface ICustomUser extends IApiUser {
    isAdmin: boolean;
    isAuthorised: boolean;
};